Version 0.8.3 Start working on map-matching. Changed the ground-truth map-matching folder to support both route and point matching. Enable the RoadNetworkGraph to support search by road and node id.
TODO: Test the Beijing point match result.

+++ src/main/java/util/io/IOService.java
@@ -43,10 +43,14 @@ public final class IOService implements Serializable {
 		File fileFolder = new File(folderPath);
 		if (fileFolder.exists() && fileFolder.listFiles() != null)
 			for (File f : Objects.requireNonNull(fileFolder.listFiles())) {
-				if (!f.delete()) try {
-					throw new IOException("ERROR! Failed to delete file.");
-				} catch (IOException e) {
-					e.printStackTrace();
+				if (f.isDirectory()) {
+					IOService.cleanFolder(f.getAbsolutePath());
+				} else {
+					if (!f.delete()) try {
+						throw new IOException("ERROR! Failed to delete file.");
+					} catch (IOException e) {
+						e.printStackTrace();
+					}
 				}
 			}
